bert-sentiment-function/vertex-ai-pipeline/scheduler-function/requirements.txt
# File: scheduler-function/requirements.txt
google-cloud-bigquery
google-cloud-aiplatform